pyproject.toml

[project]
name = ""
version = ""
description = ""
authors = [
    {name = "codest", email = "thecodest@ya.ru"},
]
dependencies = [
    "django>=4.2.2",
    "confluent-kafka>=2.1.1",
    "django-cors-headers>=4.1.0",
    "python-dotenv>=1.0.0",
]
requires-python = ">=3.10"
license = {text = "MIT"}
